backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
openai==1.30.1
redis==5.0.4
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
websockets==12.0
langdetect==1.0.9
asyncio==3.4.3
loguru==0.7.2
deepgram-sdk==3.3.0
aiohttp==3.9.5
aiosqlite==0.20.0
groq==1.0.0
google-generativeai==0.5.4
